refactor: move message-forward helpers to helpers/discord
Move isForwardedMessage and getMessageContent out of
commands/report/constructLog.ts into the lighter helpers/discord.ts
module. This keeps the spam detection pipeline's import surface minimal
— no longer pulling in date-fns, fetchSettingsEffect, and report
construction deps just for this small utility.

constructLog.ts re-exports both functions for backward compatibility.
Updated JSDoc to broaden "cross-server forwards" → "forwarded messages"
and note that snapshots may be absent/not yet loaded.
