+++ app/commands/report/constructLog.ts
@@ -1,13 +1,13 @@
 import { formatDistanceToNowStrict } from "date-fns";
-import {
-  MessageReferenceType,
-  type Message,
-  type MessageCreateOptions,
-} from "discord.js";
+import { type Message, type MessageCreateOptions } from "discord.js";
 import { Effect } from "effect";
 
 import { DiscordApiError } from "#~/effects/errors";
-import { constructDiscordLink } from "#~/helpers/discord";
+import {
+  constructDiscordLink,
+  isForwardedMessage,
+  getMessageContent,
+} from "#~/helpers/discord";
 import { truncateMessage } from "#~/helpers/string";
 import { fetchSettingsEffect, SETTINGS } from "#~/models/guilds.server";
 import { ReportReasons, type Report } from "#~/models/reportedMessages";
@@ -71,22 +71,4 @@ ${preface}
     } satisfies MessageCreateOptions;
   }).pipe(Effect.withSpan("constructLog"));
 
-export const isForwardedMessage = (message: Message): boolean => {
-  return message.reference?.type === MessageReferenceType.Forward;
-};
-
-/**
- * Returns the effective text content of a message.
- *
- * For cross-server forwards, Discord stores the original text in
- * `messageSnapshots` rather than in `message.content` (which is always "").
- * Falls back to `message.content` for non-forwarded messages or when the
- * snapshot is unexpectedly absent.
- */
-export const getMessageContent = (message: Message): string => {
-  if (isForwardedMessage(message)) {
-    const snapshot = message.messageSnapshots.first();
-    return snapshot?.content ?? message.content;
-  }
-  return message.content;
-};
+export { isForwardedMessage, getMessageContent };
