+++ app/helpers/discord.ts
@@ -1,6 +1,7 @@
 import {
   ApplicationCommandType,
   ContextMenuCommandBuilder,
+  MessageReferenceType,
   SlashCommandBuilder,
   type APIEmbed,
   type ChatInputCommandInteraction,
@@ -304,3 +305,23 @@ export function hasModRole(
   }
   return member.roles.cache.has(modRoleId);
 }
+
+export const isForwardedMessage = (message: Message): boolean => {
+  return message.reference?.type === MessageReferenceType.Forward;
+};
+
+/**
+ * Returns the effective text content of a message.
+ *
+ * For forwarded messages, Discord stores the original text in
+ * `messageSnapshots` rather than in `message.content` (which is always "").
+ * Falls back to `message.content` for non-forwarded messages or when the
+ * snapshot is unexpectedly absent or not yet loaded.
+ */
+export const getMessageContent = (message: Message): string => {
+  if (isForwardedMessage(message)) {
+    const snapshot = message.messageSnapshots.first();
+    return snapshot?.content ?? message.content;
+  }
+  return message.content;
+};
